Moved digital signature classes to a dedicated package
