+++ src/main/java/com/github/siboxd/fatturapa/model/digitalsignature/ReferenceType.java
@@ -1,4 +1,4 @@
-package com.github.siboxd.fatturapa.model;
+package com.github.siboxd.fatturapa.model.digitalsignature;
 
 import org.simpleframework.xml.Attribute;
 import org.simpleframework.xml.Element;
